feat: ui builder
