+++ lib/session/session_manager.ex
@@ -35,70 +35,72 @@ defmodule ApplicationRunner.Session.Manager do
     @on_session_stop_action
   ]
 
-  @spec send_client_event(pid(), String.t(), map()) :: :ok
-  def send_client_event(session_manager_pid, code, event) do
-    GenServer.cast(session_manager_pid, {:send_client_event, code, event})
-  end
+  # TODO: Replace by Event task
+  # @spec send_client_event(pid(), String.t(), map()) :: :ok
+  # def send_client_event(session_manager_pid, code, event) do
+  #   GenServer.cast(session_manager_pid, {:send_client_event, code, event})
+  # end
 
-  @spec start_link(keyword) :: :ignore | {:error, any} | {:ok, pid}
-  def start_link(opts) do
-    session_id = Keyword.fetch!(opts, :session_id)
-    env_id = Keyword.fetch!(opts, :env_id)
+  # @spec start_link(keyword) :: :ignore | {:error, any} | {:ok, pid}
+  # def start_link(opts) do
+  #   session_id = Keyword.fetch!(opts, :session_id)
+  #   env_id = Keyword.fetch!(opts, :env_id)
 
-    with {:ok, pid} <-
-           GenServer.start_link(__MODULE__, opts, name: {:via, :swarm, {:session, session_id}}) do
-      Swarm.join(:sessions, pid)
-      Swarm.join({:sessions, env_id}, pid)
-      {:ok, pid}
-    end
-  end
+  #   with {:ok, pid} <-
+  #          GenServer.start_link(__MODULE__, opts, name: {:via, :swarm, {:session, session_id}}) do
+  #     Swarm.join(:sessions, pid)
+  #     Swarm.join({:sessions, env_id}, pid)
+  #     {:ok, pid}
+  #   end
+  # end
 
-  @spec reload_ui(number()) :: :ok
-  def reload_ui(session_id) do
-    with {:ok, pid} <- Managers.fetch_session_manager_pid(session_id) do
-      send(pid, :data_changed)
-    end
-  end
+  # Manage by widget genserver
+  # @spec reload_ui(number()) :: :ok
+  # def reload_ui(session_id) do
+  #   with {:ok, pid} <- Managers.fetch_session_manager_pid(session_id) do
+  #     send(pid, :data_changed)
+  #   end
+  # end
 
-  @impl true
-  def init(opts) do
-    session_id = Keyword.fetch!(opts, :session_id)
-    env_id = Keyword.fetch!(opts, :env_id)
-    session_state = Keyword.fetch!(opts, :session_state)
-    user_id = Map.fetch!(session_state, :user_id)
-    function_name = Map.fetch!(session_state, :function_name)
-    assigns = Map.fetch!(session_state, :assigns)
+  # @impl true
+  # def init(opts) do
+  #   session_id = Keyword.fetch!(opts, :session_id)
+  #   env_id = Keyword.fetch!(opts, :env_id)
+  #   session_state = Keyword.fetch!(opts, :session_state)
+  #   user_id = Map.fetch!(session_state, :user_id)
+  #   function_name = Map.fetch!(session_state, :function_name)
+  #   assigns = Map.fetch!(session_state, :assigns)
 
-    {:ok, session_supervisor_pid} = Supervisor.start_link(opts)
-    # Link the process to kill the manager if the supervisor is killed.
-    # The SessionManager should be restarted by the SessionManagers then it will restart the supervisor.
-    Process.link(session_supervisor_pid)
+  #   {:ok, session_supervisor_pid} = Supervisor.start_link(opts)
+  #   # Link the process to kill the manager if the supervisor is killed.
+  #   # The SessionManager should be restarted by the SessionManagers then it will restart the supervisor.
+  #   Process.link(session_supervisor_pid)
 
-    event_handler_pid = Supervisor.fetch_module_pid!(session_supervisor_pid, EventHandler)
-    EventHandler.subscribe(event_handler_pid)
+  #   event_handler_pid = Supervisor.fetch_module_pid!(session_supervisor_pid, EventHandler)
+  #   EventHandler.subscribe(event_handler_pid)
 
-    session_state = %Session.State{
-      session_id: session_id,
-      user_id: user_id,
-      env_id: env_id,
-      function_name: function_name,
-      session_supervisor_pid: session_supervisor_pid,
-      inactivity_timeout:
-        Application.get_env(:application_runner, :session_inactivity_timeout, 1000 * 60 * 10),
-      assigns: assigns
-    }
+  #   session_state = %Session.State{
+  #     session_id: session_id,
+  #     user_id: user_id,
+  #     env_id: env_id,
+  #     function_name: function_name,
+  #     session_supervisor_pid: session_supervisor_pid,
+  #     inactivity_timeout:
+  #       Application.get_env(:application_runner, :session_inactivity_timeout, 1000 * 60 * 10),
+  #     assigns: assigns
+  #   }
 
-    first_time_user = JsonStorage.has_user_data?(env_id, user_id)
+  #   first_time_user = JsonStorage.has_user_data?(env_id, user_id)
 
-    with :ok <- create_user_data_if_needed(session_state, first_time_user),
-         :ok <- send_on_session_start_event(session_state) do
-      {:ok, session_state, session_state.inactivity_timeout}
-    else
-      {:error, reason} = err ->
-        send_error(session_state, err)
-        {:stop, reason}
-    end
-  end
+  #   with :ok <- create_user_data_if_needed(session_state, first_time_user),
+  #        :ok <- send_on_session_start_event(session_state) do
+  #     {:ok, session_state, session_state.inactivity_timeout}
+  #   else
+  #     {:error, reason} = err ->
+  #       send_error(session_state, err)
+  #       {:stop, reason}
+  #   end
+  # end
 
   defp create_user_data_if_needed(
          session_state,
